Convert main project to dotnet-csproj.
Begin the conversion process to a netstandard1.3 compatible build process, starting by converting the main project file to a dotnet new-style csproj file using the dotnet SDK. 

This means that this solution is no longer openable in Visual Studio 2013/2015. It is my intent to rectify this in a future commit, if possible, by either supplying an alternate csproj file for the original portable build, or by adding a project.json object. This presumes that VS2017 doesn't simply decide to reconvert/discard the VS2015 compatible project data whenver it opens the project.

The tooling changes allow for building both the NET40 target and the NET45+portable target from within the same project file, so the second linked-file project file for the NET40 build has been removed.

The NuSpec file is also able to be automatically generated by the new tooling, so the properties from the AssemblyInfo.cs, and the AutoMapper-NodaTime.nuspec file have been moved into the csproj file.

At the moment, this build does not include the PDB files in the main Nuget package (the default packer appears to ignore them), but a symbols+source package is created alongside the main package.

Finally, this commit adds a hastily drawn combination icon that unifies the AutoMapper icon and the NodaTime icon (pulled from their individual repositories) to use as the Nuget package icon.

At this time, all Nuget metadata URLs are pointed at the fork repository, but a branch will be created for the eventual PR back to the main repository to set the URLs to the appropriate values for that repository.